+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 4.6.3 / 2017-02-19
 * added support for tribe shortcodes
 * graceful failure if The Events Calendar is not active